Updated recursive issue
If you lets say were on week 5 day 2, and the activities were 4-7. It would try to take your week5Day1 and week5Day2 folder and throw them into the week5Day2 folder. It was really inefficient and had the user go into the zip file to remove these copied folders. Changed the line that actually looks for the activity number to make sure it includes "##-" so it would follow the current naming scheme.